Revert "Sprint task - Add uniqueness constraint and database import converter to ObjectMappingAssoziazionPO to address http://eclip.se/469940 - Part 2 - Increase importer version number" because it was causing incompatibilities
This reverts commit 6f38af4db2ab0b6ee2e8d5ad1c3275cc6285b9ee.

+++ org.eclipse.jubula.tools/src/org/eclipse/jubula/tools/internal/jarutils/IVersion.java
@@ -39,7 +39,7 @@ public interface IVersion {
      *  modifications in client code without modification of ToolkitPlugins
      *  require an increase of this version
      */
-    public final Integer JB_CLIENT_METADATA_VERSION = new Integer(7);
+    public final Integer JB_CLIENT_METADATA_VERSION = new Integer(6);
     
     /** minimum required metadata version for project import */
     public final Integer JB_CLIENT_MIN_XML_METADATA_VERSION = new Integer(5);
